Accent Reduction through Neural Style Transfer
We are following the approach of neural style transfer for speech, which is based on the idea of neural style transfer for images. Our input is a non-native speaker wav audio and our goal is to use NST to create an output wav audio with reduced accent of the speaker.